Factor out more code into modules
Decrease the size of lib.rs by moving the vector trait and error type
into separate modules.
