+++ src/lib.rs
@@ -10,217 +10,15 @@ extern crate lazy_static;
 #[macro_use]
 extern crate fixedvec;
 
+pub mod client;
 pub mod consts;
 pub mod server;
-pub mod client;
 
-pub trait VectorTrait<T: Copy> {
-    fn push(&mut self, value: T) -> Result<(), ErrorKind>;
-    fn extend(&mut self, other: &[T]) -> Result<(), ErrorKind>;
-    fn len(&self) -> usize;
-    fn is_empty(&self) -> bool;
-    fn clear(&mut self);
-    fn cut_end(&mut self, len_to_cut: usize, value: T);
-    fn as_slice(&self) -> &[T];
-    fn replace(&mut self, index: usize, value: T);
-}
+mod vector;
+pub use vector::VectorTrait;
 
-#[cfg(not(feature = "nostd"))]
-impl<T: Copy> VectorTrait<T> for Vec<T> {
-    #[inline]
-    fn push(&mut self, value: T) -> Result<(), ErrorKind> {
-        Vec::push(self, value);
-        Ok(())
-    }
-    #[inline]
-    fn extend(&mut self, values: &[T]) -> Result<(), ErrorKind> {
-        Vec::extend_from_slice(self, values);
-        Ok(())
-    }
-    #[inline]
-    fn len(&self) -> usize {
-        Vec::len(self)
-    }
-    #[inline]
-    fn is_empty(&self) -> bool {
-        Vec::is_empty(self)
-    }
-    #[inline]
-    fn clear(&mut self) {
-        Vec::clear(self);
-    }
-    fn cut_end(&mut self, len_to_cut: usize, value: T) {
-        let len = self.len();
-        if len_to_cut >= len {
-            self.clear();
-        } else {
-            self.resize(len - len_to_cut, value);
-        }
-    }
-    #[inline]
-    fn as_slice(&self) -> &[T] {
-        Vec::as_slice(self)
-    }
-    #[inline]
-    fn replace(&mut self, index: usize, value: T) {
-        self[index] = value;
-    }
-}
-
-use fixedvec::FixedVec;
-
-impl<'a, T: Copy> VectorTrait<T> for FixedVec<'a, T> {
-    #[inline]
-    fn push(&mut self, value: T) -> Result<(), ErrorKind> {
-        FixedVec::push(self, value).map_err(|_| ErrorKind::OOB)
-    }
-    #[inline]
-    fn extend(&mut self, values: &[T]) -> Result<(), ErrorKind> {
-        FixedVec::push_all(self, values).map_err(|_| ErrorKind::OOB)
-    }
-    #[inline]
-    fn len(&self) -> usize {
-        FixedVec::len(self)
-    }
-    #[inline]
-    fn is_empty(&self) -> bool {
-        FixedVec::is_empty(self)
-    }
-    #[inline]
-    fn clear(&mut self) {
-        FixedVec::clear(self);
-    }
-    fn cut_end(&mut self, len_to_cut: usize, value: T) {
-        let len = self.len();
-        if len_to_cut >= len {
-            self.clear();
-        } else {
-            self.resize(len - len_to_cut, value);
-        }
-    }
-    #[inline]
-    fn as_slice(&self) -> &[T] {
-        FixedVec::as_slice(self)
-    }
-    #[inline]
-    fn replace(&mut self, index: usize, value: T) {
-        self[index] = value;
-    }
-}
-
-#[cfg(feature = "heapless")]
-use heapless::Vec as HeaplessVec;
-
-#[cfg(feature = "heapless")]
-impl<T: Copy, const N: usize> VectorTrait<T> for HeaplessVec<T, N> {
-    #[inline]
-    fn push(&mut self, value: T) -> Result<(), ErrorKind> {
-        HeaplessVec::push(self, value).map_err(|_| ErrorKind::OOB)
-    }
-    #[inline]
-    fn extend(&mut self, values: &[T]) -> Result<(), ErrorKind> {
-        self.extend_from_slice(values).map_err(|_| ErrorKind::OOB)
-    }
-    #[inline]
-    fn len(&self) -> usize {
-        <[T]>::len(self)
-    }
-    #[inline]
-    fn is_empty(&self) -> bool {
-        HeaplessVec::is_empty(self)
-    }
-    #[inline]
-    fn clear(&mut self) {
-        HeaplessVec::clear(self);
-    }
-    fn cut_end(&mut self, len_to_cut: usize, value: T) {
-        let len = self.len();
-        if len_to_cut >= len {
-            self.clear();
-        } else {
-            self.resize(len - len_to_cut, value).unwrap();
-        }
-    }
-    #[inline]
-    fn as_slice(&self) -> &[T] {
-        HeaplessVec::as_slice(self)
-    }
-    #[inline]
-    fn replace(&mut self, index: usize, value: T) {
-        self[index] = value;
-    }
-}
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub enum ErrorKind {
-    OOB,
-    OOBContext,
-    FrameBroken,
-    FrameCRCError,
-    IllegalFunction,
-    IllegalDataAddress,
-    IllegalDataValue,
-    SlaveDeviceFailure,
-    Acknowledge,
-    SlaveDeviceBusy,
-    NegativeAcknowledge,
-    MemoryParityError,
-    GatewayPathUnavailable,
-    GatewayTargetFailed,
-    CommunicationError,
-    UnknownError,
-    Utf8Error,
-}
-
-impl ErrorKind {
-    pub fn from_modbus_error(code: u8) -> Self {
-        match code {
-            0x01 => ErrorKind::IllegalFunction,
-            0x02 => ErrorKind::IllegalDataAddress,
-            0x03 => ErrorKind::IllegalDataValue,
-            0x04 => ErrorKind::SlaveDeviceFailure,
-            0x05 => ErrorKind::Acknowledge,
-            0x06 => ErrorKind::SlaveDeviceBusy,
-            0x07 => ErrorKind::NegativeAcknowledge,
-            0x08 => ErrorKind::MemoryParityError,
-            0x09 => ErrorKind::GatewayPathUnavailable,
-            0x10 => ErrorKind::GatewayTargetFailed,
-            _ => ErrorKind::UnknownError,
-        }
-    }
-}
-
-impl core::fmt::Display for ErrorKind {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        let msg: &str = match self {
-            ErrorKind::OOB => "OUT OF BUFFER",
-            ErrorKind::OOBContext => "OUT OF BUFFER IN CONTEXT",
-            ErrorKind::FrameBroken => "FRAME BROKEN",
-            ErrorKind::FrameCRCError => "FRAME CRC ERROR",
-            ErrorKind::IllegalFunction => "MODBUS ERROR CODE 01 - ILLEGAL FUNCTION",
-            ErrorKind::IllegalDataAddress => "MODBUS ERROR CODE 02 - ILLEGAL DATA ADDRESS",
-            ErrorKind::IllegalDataValue => "MODBUS ERROR CODE 03 - ILLEGAL DATA VALUE",
-            ErrorKind::SlaveDeviceFailure => "MODBUS ERROR CODE 04 - SLAVE DEVICE FAILURE",
-            ErrorKind::Acknowledge => "MODBUS ERROR CODE 05 - ACKNOWLEDGE",
-            ErrorKind::SlaveDeviceBusy => "MODBUS ERROR CODE 06 - SLAVE DEVICE BUSY",
-            ErrorKind::NegativeAcknowledge => "MODBUS ERROR CODE 07 - NEGATIVE ACKNOWLEDGE",
-            ErrorKind::MemoryParityError => "MODBUS ERROR CODE 08 - MEMORY PARITY ERROR",
-            ErrorKind::GatewayPathUnavailable => "MODBUS ERROR CODE 10 - GATEWAY PATH UNAVAILABLE",
-            ErrorKind::GatewayTargetFailed => {
-                "MODBUS ERROR CODE 11 - GATEWAY TARGET DEVICE FAILED TO RESPOND"
-            }
-            ErrorKind::CommunicationError => {
-                "MODBUS ERROR CODE 21 - Response CRC did not match calculated CRC"
-            }
-            ErrorKind::UnknownError => "UNKNOWN MODBUS ERROR",
-            ErrorKind::Utf8Error => "UTF8 CONVERTION ERROR",
-        };
-        write!(f, "{}", msg)
-    }
-}
-
-#[cfg(not(feature = "nostd"))]
-impl std::error::Error for ErrorKind {}
+mod error;
+pub use error::ErrorKind;
 
 /// Modbus protocol selection for frame processing
 ///
